+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 * Fix crash on IN query with several thousand items.
 * Fix crash when querying indexed `NSString` properties.
 * Fixed an issue which prevented in-memory Realms from being used accross multiple threads.
+* Preserve the sort order when querying a sorted `RLMResults`.
 
 0.86.3 Release notes (2014-10-09)
 =============================================================